change all void methods to return String in order to enable javaFX to get response
